exercises/src/test/java/org/example/Exercise8Test.java
package org.example;

import org.junit.Test;

import static org.junit.Assert.assertEquals;

public class Exercise8Test {
    @Test
    public void basic() {
        /* Construct the following tree
                  1
                /   \
               /     \
              2       3
             / \     / \
            8   4   5   6
               /   / \   \
             10   7   9   5
        */

        Node root = new Node(1);
        root.left = new Node(2);
        root.right = new Node(3);
        root.left.left = new Node(8);
        root.left.right = new Node(4);
        root.right.left = new Node(5);
        root.right.right = new Node(6);
        root.left.right.left = new Node(10);
        root.right.left.left = new Node(7);
        root.right.left.right = new Node(9);
        root.right.right.right = new Node(5);

        assertEquals(18, Exercise8.findMaxSumPath(root));
    }
}
